Moved all genomic enums to participant_enums.py
