+++ rdr_service/alembic/versions/2697860e93f0_genomic_set_member_validation.py
@@ -9,7 +9,7 @@ import model.utils
 import sqlalchemy as sa
 from alembic import op
 
-from rdr_service.model.genomics import GenomicValidationFlag
+from rdr_service.participant_enums import GenomicValidationFlag
 
 # revision identifiers, used by Alembic.
 revision = "2697860e93f0"
