+++ tests/api_tests/test_genomic_set_file_handler.py
@@ -17,10 +17,8 @@ from rdr_service.model.biobank_order import BiobankOrder, BiobankOrderIdentifier
 from rdr_service.model.genomics import (
     GenomicSet,
     GenomicSetMember,
-    GenomicSetMemberStatus,
-    GenomicSetStatus,
-    GenomicValidationFlag,
 )
+from rdr_service.participant_enums import GenomicSetStatus, GenomicSetMemberStatus, GenomicValidationFlag
 from rdr_service.model.participant import Participant
 from tests import test_data
 from tests.helpers.unittest_base import BaseTestCase
