+++ tests/cron_job_tests/test_genomic_pipeline.py
@@ -27,15 +27,12 @@ from rdr_service.model.biobank_order import (
 )
 from rdr_service.model.genomics import (
     GenomicSet,
-    GenomicSetMember,
-    GenomicSetMemberStatus,
-    GenomicSetStatus,
-    GenomicSubProcessResult,
-    GenomicSubProcessStatus
+    GenomicSetMember
 )
 from rdr_service.model.participant import Participant
 from rdr_service.offline import genomic_pipeline
-from rdr_service.participant_enums import SampleStatus
+from rdr_service.participant_enums import SampleStatus, GenomicSetStatus, GenomicSetMemberStatus, \
+    GenomicSubProcessStatus, GenomicSubProcessResult
 from tests import test_data
 from tests.helpers.unittest_base import BaseTestCase
